// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/customerCreate.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomerPreferences } from './customerPreferences';
import { BillingInfo } from './billingInfo';


export interface CustomerCreate { 
    userId: string;
    billingInfo: BillingInfo;
    preferences: CustomerPreferences;
}